refactor : Id 값 Integer에서 Long으로 수정
redis에서 값을 가져올 때, byte 값에 대한 역직렬화 과정에서 convert에 byte -> Long -> Integer 형식으로 변환이 되었습니다. 과정 중 Integer로 두게될 시, 역직렬화 과정에 에러가 발생하여 Integer로 작성했던 Id값들을 Long으로 수정했습니다.
#90

+++ src/test/java/com/aliens/friendship/chatting/repository/ChattingRepositoryImplTest.java
@@ -96,7 +96,7 @@ class ChattingRepositoryImplTest {
         List<ChattingRoom> chattingRooms = new ArrayList<>();
         for (int i = 1; i <= 2; i++) {
             chattingRooms.add(chattingRoomRepository.save(ChattingRoom.builder()
-                    .id(i * (-1))
+                    .id(Long.valueOf(i * (-1)))
                     .status(ChattingRoom.RoomStatus.OPEN).build()));
         }
